Move MergedVolume to separate file
